+++ cookbooks/hostname-nativex/metadata.rb
@@ -1,8 +1,8 @@
-name             'nativex-hostname'
+name             'hostname-nativex'
 maintainer       'NativeX'
 maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
-description      'Installs/Configures nativex-hostname'
+description      'Installs/Configures hostname-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'