+++ cookbooks/hostname-nativex/recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-hostname
+# Cookbook Name:: hostname-nativex
 # Recipe:: default
 #
 # Copyright 2014, NativeX
@@ -7,4 +7,4 @@
 # All rights reserved - Do Not Redistribute
 #
 
-include_recipe "nativex-hostname::updatehostname"
+include_recipe "hostname-nativex::updatehostname"